pyproject.toml
[project]
name = "recipes"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Fricker",email = "alex@alexfricker.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "sentry-sdk[django] (>=2.46.0,<3.0.0)",
    "markdown (>=3.10,<4.0)",
    "django-filter (>=25.2,<26.0)",
    "django-constance (>=4.3.4,<5.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "ipython (>=9.8.0,<10.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.11.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
